version float

+++ include/Force.cuh
@@ -1,7 +1,7 @@
-__host__ void UpdateLog (Force *force, PlanetarySystem *sys, double *Dens, double *Energy, int TimeStep, double PhysicalTime,
+__host__ void UpdateLog (Force *force, PlanetarySystem *sys, float *Dens, float *Energy, int TimeStep, float PhysicalTime,
    int dimfxy);
 __host__ Force *AllocateForce (int dimfxy);
-__host__ void ComputeForce (Force *fc, double *Dens, double x, double y, double rsmoothing, double mass, int dimfxy, double a, double rh);
-__host__ double Compute_smoothing (double r);
+__host__ void ComputeForce (Force *fc, float *Dens, float x, float y, float rsmoothing, float mass, int dimfxy, float a, float rh);
+__host__ float Compute_smoothing (float r);
 __host__ void FreeForce (Force *force);
 __host__ void Forcescudamalloc (int dimfxy);
