+++ include/Psys.cuh
@@ -4,4 +4,4 @@ __host__ PlanetarySystem *InitPlanetarySystem (char *filename);
 __host__ PlanetarySystem *AllocPlanetSystem (int nb);
 __host__ double GetPsysInfo (PlanetarySystem *sys, int action);
 __host__ void FreePlanetary (PlanetarySystem *sys);
-__host__ void RotatePsys (PlanetarySystem *sys, double angle);
+__host__ void RotatePsys (PlanetarySystem *sys, float angle);
